Merged PR 11: Encryption and decryption testing is done
Encryption and decryption testing is done

The test runner has been modified to point to either the encryption or decryption feature only. Some tests fail as expected.  Still trying to find a method for using Selenium to test the clipboard for the copy feature, right now that test is being automatically failed because we know the feature is broken. It will be implemented when that bug fix occurs